Remove Schedule complicated logic.
It's easier if we model a Schedule as a sequence of all tasks, cutting
off the tasks that we would schedule past the deadline. Then we can
represent any transformation in the schedule as a swap of two tasks
(instead of having to randomly choose between a swap, insert, and
remove).